[feat]: crud pacientes, psicologos e atendimentos  -  [fix]: bug de atuaizacao e autorizacao

+++ src/controlls/psicologosController.js
@@ -1,79 +1,79 @@
 import { Psicologos } from "../models/index.js";
-
+import bcrypt from 'bcryptjs';
 
 const psicologosController = {
 
     listarPsicologos: async (req, res) =>{
 
         const listaPsicologos = await Psicologos.findAll();
-        res.json(listaPsicologos);
+        res.status(200).json(listaPsicologos);
     },
 //
     listarPsicologo: async (req, res) => {
-        const listaPsicologos = req.params;
-        const psicologo = await Psicologos.findByPk(id, {
-         
-        });
+        const {id} = req.params;
+        const psicologo = await Psicologos.findOne({
+            where: {
+                id,
+            },
+        })
+
         if (psicologo) {
           res.json(psicologo);
         } else {
-          res.status(404).json({ message: "Psicologo not found" });
+          res.status(404).json({ message: "Id não encontrado" });
         }
       },
 //
     async cadastrarPsicologos(req, res){
         
         const { nome, email, senha, apresentacao  } = req.body;
+        const novaSenha = bcrypt.hashSync(senha, 10);
         const novoPsicologo = await Psicologos.create({
             nome,
             email,
-            senha,
-            apresentacao
-
+            senha: novaSenha,
+            apresentacao,
         });
 
         res.status(201).json(novoPsicologo);
     },
 //
     async atualizarPsicologo(req,res){
-        const {id} = req.params;
+        const { id } = req.params;
         const { nome, email, senha, apresentacao } = req.body;
+        const novaSenha = bcrypt.hashSync(senha, 10);
 
-        const psicologoAtualizado = await Paceintes.update({
+        const [psicologoAtualizado] = await Psicologos.update({
             nome, 
             email, 
-            senha,
+            senha: novaSenha,
             apresentacao
         },
         {
-            where: {
-                id,
-            }
+            where: { id } , 
+           
         }
+        
 
         );
-        res.json('Psicologo Atualizado')
+        const psicologo = await Psicologos.findOne({ where: { id } });
+        res.status(200).json(psicologo);
     
     },
 //
-    async deletarPaciente(req, res){
-        try { // tenta fazer o comando abaixo
-            const {id} = req.params;
-
+ async deletarPsicologo(req, res){
+        const { id } = req.params;
+        const deletaPsicologo = await Psicologos.findByPk(id);
+        if(!deletaPsicologo){
+            return res.status(404).json('Id não encontrado');
+        }
         await Psicologos.destroy({
-            where: {
-                id,
-            },
+            where: { id}
         });
 
-        res.status(204);
+        res.status(204).send();
 
-        }catch(error){ // trata o erro e especifica o que fazer em caso de erro
-            return res.status(500).json('Ocorreu algum problema')
-        }
-        
-    },
-
+    } 
 };
 
 export default psicologosController;
